+++ services/backend/src/main/java/com/soprasteria/g4it/backend/apidigitalservice/modeldb/DigitalService.java
@@ -49,6 +49,11 @@ public class DigitalService {
     private String name;
 
     /**
+     * Specifies whether it's an IA service
+     */
+    private boolean isAi;
+
+    /**
      * The Criterias key.
      */
     private List<String> criteria;