+++ services/backend/src/main/java/com/soprasteria/g4it/backend/apievaluating/business/EvaluatingService.java
@@ -162,6 +162,7 @@ public class EvaluatingService {
                 .datetime(LocalDateTime.now())
                 .hasVirtualEquipments(true)
                 .hasApplications(false)
+                .isAi(digitalService.isAi())
                 .build();
 
         List<String> activeCriteria = criteriaService.getSelectedCriteriaForDigitalService(subscriber, organizationId, digitalService.getCriteria())