+++ services/backend/src/main/java/com/soprasteria/g4it/backend/apiinout/mapper/InputToCsvRecord.java
@@ -8,10 +8,12 @@
 
 package com.soprasteria.g4it.backend.apiinout.mapper;
 
+import com.soprasteria.g4it.backend.apiaiinfra.modeldb.InAiInfrastructure;
 import com.soprasteria.g4it.backend.apiinout.modeldb.InApplication;
 import com.soprasteria.g4it.backend.apiinout.modeldb.InDatacenter;
 import com.soprasteria.g4it.backend.apiinout.modeldb.InPhysicalEquipment;
 import com.soprasteria.g4it.backend.apiinout.modeldb.InVirtualEquipment;
+import com.soprasteria.g4it.backend.apiparameterai.modeldb.InAiParameter;
 import org.mapstruct.Mapper;
 
 import java.util.List;
@@ -29,6 +31,31 @@ public interface InputToCsvRecord {
                 datacenter.getLocation(),
                 "", // entityName
                 "" // source
+        );
+    }
+
+    default List<String> toCsv(InAiParameter inAiParameter) {
+        return List.of(
+                inAiParameter.getDigitalServiceUid(),
+                inAiParameter.getModelName(),
+                inAiParameter.getNbParameters(),
+                inAiParameter.getFramework(),
+                inAiParameter.getQuantization(),
+                inAiParameter.getTotalGeneratedTokens().toString(),
+                inAiParameter.getNumberUserYear().toString(),
+                inAiParameter.getAverageNumberRequest().toString(),
+                inAiParameter.getAverageNumberToken().toString(),
+                inAiParameter.getIsInference().toString(),
+                inAiParameter.getIsFinetuning().toString()
+        );
+    }
+
+    default List<String> toCsv(InAiInfrastructure inAiInfrastructure) {
+        return List.of(
+                inAiInfrastructure.getDigitalServiceUid(),
+                inAiInfrastructure.getComplementaryPue().toString(),
+                inAiInfrastructure.getNbGpu().toString(),
+                inAiInfrastructure.getGpuMemory().toString()
         );
     }