+++ services/frontend/src/app/layout/digital-services-footprint/digital-services-footprint-dashboard/bar-chart/bar-chart.component.ts
@@ -20,7 +20,7 @@ import {
     DigitalServiceServersImpact,
     DigitalServiceTerminalsImpact,
     ImpactACVStep,
-    ImpactNetworkSipValue,
+    ImpactNetworkNames,
     ImpactSipValue,
     ServerImpact,
     ServersType,
@@ -135,40 +135,41 @@ export class BarChartComponent extends AbstractDashboard {
         );
         let xAxis: any[] = [];
         const yAxis: any[] = [];
-        seriesData.forEach((impact: ImpactNetworkSipValue) => {
+        seriesData.forEach((impact: ImpactNetworkNames) => {
             networkMap[impact.networkType] = {
                 ...impact,
-                rawValue:
-                    (networkMap[impact.networkType]?.rawValue ?? 0) + impact.rawValue,
-                sipValue:
-                    (networkMap[impact.networkType]?.sipValue ?? 0) + impact.sipValue,
                 status: {
-                    ok:
-                        (networkMap[impact.networkType]?.status?.ok ?? 0) +
-                        (impact.status === Constants.DATA_QUALITY_STATUS.ok
-                            ? impact.countValue
-                            : 0),
-                    error:
-                        (networkMap[impact.networkType]?.status?.error ?? 0) +
-                        (impact.status !== Constants.DATA_QUALITY_STATUS.ok
-                            ? impact.countValue
-                            : 0),
-                    total:
-                        (networkMap[impact.networkType]?.status?.total ?? 0) +
-                        impact.countValue,
+                    ok: impact.status.ok,
+                    error: impact.status.error,
+                    total: impact.status.total,
                 },
             };
-        });
-        xAxis = Object.keys(networkMap);
-        xAxis.forEach((key) => {
-            const impact = networkMap[key];
-            yAxis.push({
-                value: impact.sipValue < 1 ? impact.sipValue : impact.sipValue.toFixed(0),
-                rawValue: impact.rawValue,
-                unit: impact.unit,
+            xAxis.push(impact.networkType);
+            impact.items.forEach((item: any, index: number) => {
+                const data = [
+                    impact.networkType,
+                    item.sipValue < 1 ? item.sipValue : item.sipValue.toFixed(0),
+                    item.rawValue,
+                    item.unit,
+                ];
+
+                yAxis.push({
+                    name: item.name,
+                    data: [data],
+                    type: "bar",
+                    stack: "Ad",
+                    emphasis: {
+                        focus: "series",
+                    },
+                    itemStyle: {
+                        color: this.createStackBarGradientColor(
+                            index,
+                            impact.items.length,
+                        ),
+                    },
+                });
             });
         });
-
         this.xAxisInput = xAxis;
         this.criteriaMap = networkMap;
         return {
@@ -176,10 +177,14 @@ export class BarChartComponent extends AbstractDashboard {
                 show: true,
                 formatter: (params: any) => {
                     return `
+                    <div style="display: flex; align-items: center; height: 30px;">
+                            <span style="display: inline-block; width: 10px; height: 10px; background-color: ${params.color}; border-radius: 50%; margin-right: 5px;"></span>
+                            <span style="font-weight: bold; margin-right: 15px;">${params.seriesName}</span>
+                        </div>
                         <div>
-                            ${this.integerPipe.transform(params.value)}
+                            ${this.integerPipe.transform(params.data[1])}
                             ${this.translate.instant("common.peopleeq-min")}<br>
-                            ${this.decimalsPipe.transform(params.data.rawValue)} ${params.data.unit}
+                            ${this.decimalsPipe.transform(params.data[2])} ${params.data[3]}
                         </div>
                     `;
                 },
@@ -215,13 +220,7 @@ export class BarChartComponent extends AbstractDashboard {
                     type: "value",
                 },
             ],
-            series: [
-                {
-                    name: "networks",
-                    type: "bar",
-                    data: yAxis,
-                },
-            ],
+            series: yAxis,
             color: Constants.BLUE_COLOR,
         };
     }
@@ -293,27 +292,39 @@ export class BarChartComponent extends AbstractDashboard {
                     type: "value",
                 },
             ],
-            series: [
-                {
-                    name: isTerminal ? "terminals" : "Cloud Services",
-                    type: "bar",
-                    data: yAxis,
-                },
-            ],
+            series: this.barChartChild
+                ? [
+                      {
+                          name: isTerminal ? "terminals" : "Cloud Services",
+                          type: "bar",
+                          data: yAxis,
+                      },
+                  ]
+                : yAxis,
             color: Constants.BLUE_COLOR,
         } as any;
     }
 
     private createTooltipFormatter(isTerminal: boolean): (params: any) => string {
         return (params: any) => {
-            if (params.value === undefined) {
+            if (
+                this.barChartChild
+                    ? params.value === undefined
+                    : params.data[1] === undefined
+            ) {
                 return "";
             }
+            const seriesName = !this.barChartChild
+                ? `<div style="display: flex; align-items: center; height: 30px;">
+                            <span style="display: inline-block; width: 10px; height: 10px; background-color: ${params.color}; border-radius: 50%; margin-right: 5px;"></span>
+                            <span style="font-weight: bold; margin-right: 15px;">${params.seriesName}</span>
+                        </div>`
+                : "";
             const showedHtml = `
                 <div>
-                    ${this.integerPipe.transform(params.value)}
+                    ${this.integerPipe.transform(this.barChartChild ? params.value : params.data[1])}
                     ${this.translate.instant("common.peopleeq-min")} <br>
-                    ${this.decimalsPipe.transform(params.data.rawValue)} ${params.data.unit}
+                    ${this.decimalsPipe.transform(params.data.rawValue ?? params.data[4])} ${params.data.unit ?? params.data[5]}
                 </div>
             `;
 
@@ -326,20 +337,20 @@ export class BarChartComponent extends AbstractDashboard {
                             isTerminal
                                 ? "digital-services-terminals.nb-user"
                                 : "digital-services-cloud-services.tooltip_average_workload",
-                        )}: ${this.decimalsPipe.transform(params.data[isTerminal ? "nbUsers" : "averageWorkLoad"])}${isTerminal ? "" : "%"}
+                        )}: ${this.decimalsPipe.transform(params.data[2])}${isTerminal ? "" : "%"}
                     </div>
                     <div>
                         ${this.translate.instant(
                             isTerminal
                                 ? "digital-services-terminals.yearly-usage"
                                 : "digital-services-cloud-services.tooltip_annual_usage",
-                        )}: ${this.decimalsPipe.transform(params.data[isTerminal ? "usageTime" : "averageUsage"])}
+                        )}: ${this.decimalsPipe.transform(params.data[3])}
                         ${this.translate.instant("digital-services-terminals.hours")}
                     </div>
                 `;
             }
 
-            return showedHtml + otherHtml;
+            return seriesName + showedHtml + otherHtml;
         };
     }
 
@@ -418,38 +429,39 @@ export class BarChartComponent extends AbstractDashboard {
         seriesData.forEach((impact: any) => {
             okMap[impact.name] = {
                 status: {
-                    ok: impact.impact.reduce(
-                        (acc: number, i: any) => acc + (i.statusCount?.ok ?? 0),
-                        0,
-                    ),
-                    error: impact.impact.reduce(
-                        (acc: number, i: any) => acc + (i.statusCount?.error ?? 0),
-                        0,
-                    ),
-                    total: impact.impact.reduce(
-                        (acc: number, i: any) => acc + (i.statusCount?.total ?? 0),
-                        0,
-                    ),
+                    ok: impact.status.ok ?? 0,
+                    error: impact.status.error ?? 0,
+                    total: impact.status.total ?? 0,
                 },
             };
             xAxis.push(impact.name);
-            yAxis.push({
-                value:
-                    impact.totalSipValue < 1
-                        ? impact.totalSipValue
-                        : impact.totalSipValue.toFixed(0),
-                name: impact.name,
-                ...(isTerminals
-                    ? {
-                          nbUsers: impact.totalNbUsers,
-                          usageTime: impact.avgUsageTime,
-                      }
-                    : {
-                          averageUsage: impact.totalAvgUsage,
-                          averageWorkLoad: impact.totalAvgWorkLoad,
-                      }),
-                rawValue: impact.rawValue,
-                unit: impact.unit,
+
+            const items = isTerminals ? impact.terminals : impact.clouds;
+            items.forEach((item: any, index: number) => {
+                const data = [
+                    impact.name,
+                    item.totalSipValue < 1
+                        ? item.totalSipValue
+                        : item.totalSipValue.toFixed(0),
+                    ...(isTerminals
+                        ? [item.totalNbUsers, item.avgUsageTime]
+                        : [item.totalAvgWorkLoad, item.totalAvgUsage]),
+                    item.rawValue,
+                    item.unit,
+                ];
+
+                yAxis.push({
+                    name: item.name,
+                    data: [data],
+                    type: "bar",
+                    stack: "Ad",
+                    emphasis: {
+                        focus: "series",
+                    },
+                    itemStyle: {
+                        color: this.createStackBarGradientColor(index, items.length),
+                    },
+                });
             });
         });
         return okMap;
@@ -462,9 +474,16 @@ export class BarChartComponent extends AbstractDashboard {
         okMap: StatusCountMap,
         isTerminals: boolean,
     ): StatusCountMap {
-        const childData = seriesData.find(
-            (item: any) => item.name === this.selectedDetailParam,
-        );
+        let childData;
+        if (!isTerminals) {
+            childData = seriesData
+                .find((item: any) => item.name === this.selectedDetailParam)
+                .clouds.find((c: any) => c.name === this.selectedDetailName);
+        } else {
+            childData = seriesData
+                .find((item: any) => item.name === this.selectedDetailParam)
+                .terminals.find((term: any) => term.name === this.selectedDetailName);
+        }
 
         const stepKey = "acvStep";
 
@@ -857,17 +876,5 @@ export class BarChartComponent extends AbstractDashboard {
         const g = Math.round((1 - t) * startG + t * endG);
         const b = Math.round((1 - t) * startB + t * endB);
         return `rgb(${r},${g},${b})`;
-    }
-
-    selectedStackBarClick(event: string): void {
-        if (
-            (this.selectedParam == "Terminal" ||
-                this.selectedParam === Constants.CLOUD_SERVICE) &&
-            !this.barChartChild
-        ) {
-            this.barChartChildChange.emit(true);
-            this.selectedDetailNameChange.emit("terminals");
-            this.selectedDetailParamChange.emit(event);
-        }
     }
 }