+++ services/frontend/src/app/layout/digital-services-footprint/digital-services-footprint.module.ts
@@ -25,6 +25,8 @@ import { DividerModule } from "primeng/divider";
 import { ScrollPanelModule } from "primeng/scrollpanel";
 import { SharedModule } from "src/app/core/shared/shared.module";
 import { DigitalServiceTableComponent } from "../common/digital-service-table/digital-service-table.component";
+import { DigitalServicesAiInfrastructureComponent } from "./digital-services-ai-infrastructure/digital-services-ai-infrastructure.component";
+import { DigitalServicesAiParametersComponent } from "./digital-services-ai-parameters/digital-services-ai-parameters.component";
 import { DigitalServicesCloudServicesSidePanelComponent } from "./digital-services-cloud-services/digital-services-cloud-services-side-panel/digital-services-cloud-services-side-panel.component";
 import { DigitalServicesCloudServicesComponent } from "./digital-services-cloud-services/digital-services-cloud-services.component";
 import { BarChartComponent } from "./digital-services-footprint-dashboard/bar-chart/bar-chart.component";
@@ -68,6 +70,8 @@ import { DigitalServicesTerminalsComponent } from "./digital-services-terminals/
         PanelServerParametersComponent,
         PanelAddVmComponent,
         PanelListVmComponent,
+        DigitalServicesAiParametersComponent,
+        DigitalServicesAiInfrastructureComponent,
     ],
     imports: [
         CommonModule,
@@ -92,6 +96,7 @@ import { DigitalServicesTerminalsComponent } from "./digital-services-terminals/
         RadioButtonModule,
         DividerModule,
         digitalServicesFootprintRouter,
+        TableModule,
     ],
     schemas: [CUSTOM_ELEMENTS_SCHEMA, NO_ERRORS_SCHEMA],
     exports: [DigitalServicesFootprintComponent],