+++ services/frontend/src/app/layout/welcome-page/welcome-page.component.ts
@@ -17,7 +17,7 @@ import { firstValueFrom } from "rxjs";
 import { Subscriber } from "src/app/core/interfaces/administration.interfaces";
 import { UserService } from "src/app/core/service/business/user.service";
 import { WorkspaceService } from "src/app/core/service/business/workspace.service";
-
+import { environment } from "src/environments/environment";
 @Component({
     selector: "app-welcome-page",
     templateUrl: "./welcome-page.component.html",
@@ -38,10 +38,31 @@ export class WelcomePageComponent {
     currentSubscriber: Subscriber = {} as Subscriber;
     isAllowedInventory: boolean = false;
     isAllowedDigitalService: boolean = false;
+    isAllowedEcoMindAi: boolean = false;
+    isEcoMindEnabledForCurrentSubscriber: boolean = false;
+    isEcoMindModuleEnabled: boolean = environment.isEcomindEnabled;
 
     private readonly destroyRef = inject(DestroyRef);
     public userService = inject(UserService);
     ecoDesignPercent = this.userService.ecoDesignPercent;
+
+    externalLinks = [
+        {
+            href: "https://github.com/G4ITTeam/g4it",
+            label: "common.github-link",
+            iconClass: "pi pi-sign-out ml-auto",
+        },
+        {
+            href: "https://saas-g4it.com/documentation/",
+            label: "common.doc-link",
+            iconClass: "pi pi-sign-out ml-auto",
+        },
+        {
+            href: "https://github.com/Boavizta/boaviztapi",
+            label: "common.boaviztapi-github-link",
+            iconClass: "pi pi-sign-out ml-3 lg:ml-4",
+        },
+    ];
 
     constructor(
         private workspaceService: WorkspaceService,
@@ -60,11 +81,17 @@ export class WelcomePageComponent {
             .subscribe((isAllowed: boolean) => {
                 this.isAllowedDigitalService = isAllowed;
             });
+        this.userService.isAllowedEcoMindAiRead$
+            .pipe(takeUntilDestroyed(this.destroyRef))
+            .subscribe((isAllowed: boolean) => {
+                this.isAllowedEcoMindAi = isAllowed;
+            });
         this.userName = userDetails?.firstName + " " + userDetails?.lastName;
 
-        this.userService.currentSubscriber$.subscribe(
-            (subscriber: any) => (this.currentSubscriber = subscriber),
-        );
+        this.userService.currentSubscriber$.subscribe((subscriber: any) => {
+            this.currentSubscriber = subscriber;
+            this.isEcoMindEnabledForCurrentSubscriber = this.currentSubscriber.ecomindai;
+        });
 
         this.userService.currentOrganization$
             .pipe(takeUntilDestroyed(this.destroyRef))
@@ -87,7 +114,19 @@ export class WelcomePageComponent {
 
     digitalServices() {
         if (this.isAllowedDigitalService) {
-            this.router.navigateByUrl(`${this.selectedPath}/digital-services`);
+            this.router.navigateByUrl(`${this.selectedPath}/digital-services`, {
+                state: { isIa: false },
+            });
+        } else {
+            this.router.navigateByUrl("/useful-information");
+        }
+    }
+
+    ecoMindAi() {
+        if (this.isAllowedEcoMindAi) {
+            this.router.navigateByUrl(`${this.selectedPath}/eco-mind-ai`, {
+                state: { isIa: true },
+            });
         } else {
             this.router.navigateByUrl("/useful-information");
         }